+++ src/steps/resource-manager/storage/client.test.ts
@@ -5,7 +5,6 @@ import {
   StorageQueue,
   Table,
   Kind,
-  BlobServiceProperties,
 } from '@azure/arm-storage/esm/models';
 import {
   createMockIntegrationLogger,
@@ -15,7 +14,7 @@ import {
 import config, {
   configFromEnv,
 } from '../../../../test/integrationInstanceConfig';
-import { StorageClient } from './client';
+import { StorageClient, createStorageAccountServiceClient } from './client';
 import { IntegrationConfig } from '../../../types';
 import {
   setupAzureRecording,
@@ -64,37 +63,102 @@ describe('iterateStorageAccounts', () => {
   });
 });
 
-describe('getBlobServiceProperties', () => {
-  test('all', async () => {
+describe('createStorageAccountServiceClient', () => {
+  async function getSetupData() {
     recording = setupAzureRecording({
       directory: __dirname,
-      name: 'getBlobServiceProperties',
+      name: 'storageAccountServiceClient-getSetupData',
       options: {
         matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
       },
     });
-
     const client = new StorageClient(
       configFromEnv,
       createMockIntegrationLogger(),
     );
 
-    const b: BlobServiceProperties[] = [];
-    await client.iterateStorageAccounts(async (sa) => {
-      b.push(
-        await client.getBlobServiceProperties({ name: sa.name!, id: sa.id! }),
-      );
+    const storageAccounts: StorageAccount[] = [];
+    await client.iterateStorageAccounts((sa) => {
+      storageAccounts.push(sa);
     });
 
-    expect(b).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          deleteRetentionPolicy: {
-            enabled: expect.any(Boolean),
-          },
-        }),
-      ]),
+    const j1devStorageAccounts = storageAccounts.filter((sa) =>
+      sa.name?.endsWith('j1dev'),
     );
+    expect(j1devStorageAccounts.length).toBe(1);
+    const storageAccount = j1devStorageAccounts[0];
+
+    await recording.stop();
+    return { storageAccount };
+  }
+
+  describe('getBlobServiceProperties', () => {
+    test('success', async () => {
+      const { storageAccount } = await getSetupData();
+
+      recording = setupAzureRecording({
+        directory: __dirname,
+        name: 'storageAccountServiceClient-getBlobServiceProperties',
+        options: {
+          matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+        },
+      });
+
+      const storageAccountServiceClient = createStorageAccountServiceClient({
+        config: configFromEnv,
+        logger: createMockIntegrationLogger(),
+        storageAccount: {
+          name: storageAccount.name!,
+          kind: storageAccount.kind!,
+        },
+      });
+
+      const response = await storageAccountServiceClient.getBlobServiceProperties();
+
+      expect(response).toMatchObject({
+        blobAnalyticsLogging: {
+          read: expect.any(Boolean),
+          write: expect.any(Boolean),
+          deleteProperty: expect.any(Boolean),
+        },
+        deleteRetentionPolicy: {
+          enabled: expect.any(Boolean),
+        },
+      });
+    });
+  });
+
+  describe('getQueueServiceProperties', () => {
+    test('success', async () => {
+      const { storageAccount } = await getSetupData();
+
+      recording = setupAzureRecording({
+        directory: __dirname,
+        name: 'storageAccountServiceClient-getQueueServiceProperties',
+        options: {
+          matchRequestsBy: getMatchRequestsBy({ config: configFromEnv }),
+        },
+      });
+
+      const storageAccountServiceClient = createStorageAccountServiceClient({
+        config: configFromEnv,
+        logger: createMockIntegrationLogger(),
+        storageAccount: {
+          name: storageAccount.name!,
+          kind: storageAccount.kind!,
+        },
+      });
+
+      const response = await storageAccountServiceClient.getQueueServiceProperties();
+
+      expect(response).toMatchObject({
+        queueAnalyticsLogging: {
+          read: expect.any(Boolean),
+          write: expect.any(Boolean),
+          deleteProperty: expect.any(Boolean),
+        },
+      });
+    });
   });
 });