+++ src/steps/resource-manager/storage/client.ts
@@ -1,4 +1,6 @@
 import { StorageManagementClient } from '@azure/arm-storage';
+import { BlobServiceClient } from '@azure/storage-blob';
+import { QueueServiceClient } from '@azure/storage-queue';
 import {
   BlobContainer,
   FileShare,
@@ -6,7 +8,6 @@ import {
   Table,
   StorageQueue,
   Kind,
-  BlobServicesGetServicePropertiesResponse,
 } from '@azure/arm-storage/esm/models';
 
 import {
@@ -14,6 +15,44 @@ import {
   iterateAllResources,
 } from '../../../azure/resource-manager/client';
 import { resourceGroupName } from '../../../azure/utils';
+import { IntegrationConfig } from '../../../types';
+import { IntegrationLogger } from '@jupiterone/integration-sdk-core';
+import { ClientSecretCredential } from '@azure/identity';
+
+export function createStorageAccountServiceClient(options: {
+  config: IntegrationConfig;
+  logger: IntegrationLogger;
+  storageAccount: { name: string; kind: Kind };
+}) {
+  const { config, storageAccount } = options;
+  const credential = new ClientSecretCredential(
+    config.directoryId,
+    config.clientId,
+    config.clientSecret,
+  );
+
+  return {
+    getBlobServiceProperties: async () => {
+      const client = new BlobServiceClient(
+        `https://${storageAccount.name}.blob.core.windows.net`,
+        credential,
+      );
+      const response = await client.getProperties();
+      return response._response.parsedBody;
+    },
+
+    getQueueServiceProperties: async () => {
+      if (isServiceEnabledForKind.queue(storageAccount.kind)) {
+        const client = new QueueServiceClient(
+          `https://${storageAccount.name}.queue.core.windows.net`,
+          credential,
+        );
+        const response = await client.getProperties();
+        return response._response.parsedBody;
+      }
+    },
+  };
+}
 
 export class StorageClient extends Client {
   public async iterateStorageAccounts(
@@ -29,21 +68,6 @@ export class StorageClient extends Client {
       resourceDescription: 'storage.storageAccounts',
       callback,
     });
-  }
-
-  public async getBlobServiceProperties(storageAccount: {
-    name: string;
-    id: string;
-  }): Promise<BlobServicesGetServicePropertiesResponse> {
-    const serviceClient = await this.getAuthenticatedServiceClient(
-      StorageManagementClient,
-    );
-    const resourceGroup = resourceGroupName(storageAccount.id, true)!;
-    const accountName = storageAccount.name;
-    return serviceClient.blobServices.getServiceProperties(
-      resourceGroup,
-      accountName,
-    );
   }
 
   public async iterateStorageBlobContainers(
@@ -88,7 +112,7 @@ export class StorageClient extends Client {
     const resourceGroup = resourceGroupName(storageAccount.id, true)!;
     const accountName = storageAccount.name!;
 
-    if ((['Storage', 'StorageV2'] as Kind[]).includes(storageAccount.kind)) {
+    if (isServiceEnabledForKind.queue(storageAccount.kind)) {
       try {
         await iterateAllResources({
           logger: this.logger,
@@ -137,7 +161,7 @@ export class StorageClient extends Client {
     const resourceGroup = resourceGroupName(storageAccount.id, true)!;
     const accountName = storageAccount.name;
 
-    if ((['Storage', 'StorageV2'] as Kind[]).includes(storageAccount.kind)) {
+    if (isServiceEnabledForKind.table(storageAccount.kind)) {
       try {
         await iterateAllResources({
           logger: this.logger,
@@ -206,3 +230,12 @@ export class StorageClient extends Client {
     });
   }
 }
+
+const isServiceEnabledForKind = {
+  table: (kind: Kind) => {
+    return (['Storage', 'StorageV2'] as Kind[]).includes(kind);
+  },
+  queue: (kind: Kind) => {
+    return (['Storage', 'StorageV2'] as Kind[]).includes(kind);
+  },
+};