+++ src/steps/resource-manager/storage/index.ts
@@ -10,7 +10,7 @@ import {
 import { createAzureWebLinker } from '../../../azure';
 import { IntegrationStepContext, IntegrationConfig } from '../../../types';
 import { ACCOUNT_ENTITY_TYPE, STEP_AD_ACCOUNT } from '../../active-directory';
-import { StorageClient } from './client';
+import { StorageClient, createStorageAccountServiceClient } from './client';
 import { steps, entities, relationships } from './constants';
 import {
   createStorageContainerEntity,
@@ -33,6 +33,24 @@ export async function fetchStorageAccounts(
   executionContext: IntegrationStepContext,
 ): Promise<void> {
   const { instance, logger, jobState } = executionContext;
+  async function getStorageAccountServiceProperties(storageAccount: {
+    name: string;
+    kind: Kind;
+  }) {
+    const storageAccountServiceClient = createStorageAccountServiceClient({
+      config: instance.config,
+      logger,
+      storageAccount,
+    });
+
+    const storageBlobServiceProperties = await storageAccountServiceClient.getBlobServiceProperties();
+    const storageQueueServiceProperties = await storageAccountServiceClient.getQueueServiceProperties();
+
+    return {
+      blob: storageBlobServiceProperties,
+      queue: storageQueueServiceProperties,
+    };
+  }
   const client = new StorageClient(instance.config, logger);
 
   const accountEntity = await jobState.getData<Entity>(ACCOUNT_ENTITY_TYPE);
@@ -41,15 +59,14 @@ export async function fetchStorageAccounts(
   const keyVaultEntityMap = await buildKeyVaultEntityMap(executionContext);
 
   await client.iterateStorageAccounts(async (storageAccount) => {
-    const storageBlobServiceProperties = await client.getBlobServiceProperties({
-      name: storageAccount.name!,
-      id: storageAccount.id!,
-    });
+    const storageAccountServiceProperties = await getStorageAccountServiceProperties(
+      { name: storageAccount.name!, kind: storageAccount.kind! },
+    );
     const storageAccountEntity = await jobState.addEntity(
       createStorageAccountEntity(
         webLinker,
         storageAccount,
-        storageBlobServiceProperties,
+        storageAccountServiceProperties,
       ),
     );
     await createResourceGroupResourceRelationship(